conteiner/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
celery
requests
aiohttp
dnspython
passlib
python-jose[cryptography]
httpx
pytest
pytest-asyncio
pytest-cov
email-validator
argon2-cffi
flower
python-dotenv
ping3
redis
paramiko